% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alogistic.R
\name{alogistic}
\alias{alogistic}
\title{Logistic growth with the Allee effect}
\usage{
alogistic(t, y, p)
}
\arguments{
\item{t}{Time points for which N wll be returned.}

\item{y}{N, population size}

\item{p}{a vector of logistic growth (r, alpha) and Allee effect parameters (a, tau); a is the threshold population size.}
}
\description{
A function for continuous logistic growth with the Allee effect, for use with \code{ode} in the \code{deSolve} package.
}
\details{
The user does not put these directly into this function, but rather uses \code{ode} in the \code{deSolve} package.
The function is based on the logistic growth equation
\deqn{\frac{dN}{dt} = rN\left(1-\alpha N\right)\left(1 - \frac{a+\tau}{N+\tau}\right)}
with \eqn{a} being the threshold population size.
}
\examples{
library(deSolve)
p <- c(r=1,alpha=.01, a=10, tau=.1)
time <- 0:10
initialN <- 11
out <- ode(y=initialN, times=time,
             func=alogistic, parms=p)
plot(time, out[,-1], type='l')

}
\references{
Stevens. M.H.H. (2009) \emph{A Primer of Ecology with R}. Use R! Series. Springer.
}
\seealso{
\code{\link{clogistic}},\code{\link{dlogistic}}, \code{\link{thetalogistic}}, \code{\link{levins}}
}
\author{
Hank Stevens
}
\keyword{Allee}
\keyword{effect}
\keyword{logistic}
