% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collatz_sequence.R
\name{collatz_sequence}
\alias{collatz_sequence}
\title{Generate the Collatz Sequence}
\usage{
collatz_sequence(n)
}
\arguments{
\item{n}{The starting number for the sequence.}
}
\value{
A numeric vector representing the Collatz sequence starting from the given number.
}
\description{
Generate the Collatz Sequence
}
\examples{
collatz_sequence(6) # 6, 3, 10, 5, 16, 8, 4, 2, 1
}
