% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer-package.R
\docType{data}
\name{coneflower}
\alias{coneflower}
\title{Smooth coneflower data}
\format{
A data frame with 136 rows and 5 variables: \describe{
\item{logA}{natural log of total leaf area in the first year}
\item{logAp}{natural log of total leaf area in the second year ('A prime')}
\item{surv}{indicator of survival from year one to year 2 (0,1)}
\item{flower_p}{indicator of flowering in year 2 (0,1)} }
}
\source{

}
\description{
A dataset containing size, survival, and flowering in smooth coneflower.
}
\keyword{datasets}
