% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer-package.R
\docType{data}
\name{coneflowerrecruits}
\alias{coneflowerrecruits}
\title{Smooth coneflower data}
\format{
A data frame with 36 rows and 3 variables: \describe{
\item{area17}{total leaf area in the first year} \item{surv18}{indicator of
survival from year one to year 2 (0,1)} \item{area19}{total leaf area in the
third year} }
}
\source{

}
\description{
A dataset containing recruit size and survival.
}
\keyword{datasets}
