% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coneflowerseeds.R
\docType{data}
\name{coneflowerseeds}
\alias{coneflowerseeds}
\title{Smooth coneflower seed data}
\format{
A data frame with 136 rows and 5 variables:
\describe{
  \item{logAs}{natural log of total leaf area in the first year}
  \item{seeds}{number of seeds produced in the first year}
}
}
\source{

}
\usage{
coneflowerseeds
}
\description{
A dataset containing size and seed set of individual coneflowers.
}
\keyword{datasets}
