% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlogistic.R
\name{dlogistic}
\alias{dlogistic}
\title{Discrete Logistic Growth}
\usage{
dlogistic(alpha = 0.01, rd = 1, N0 = 2, t = 15)
}
\arguments{
\item{alpha}{per capita negative density dependence (a positive value will
result in the usual negative effect)}

\item{rd}{discrete growth increment}

\item{N0}{initial population size}

\item{t}{end time point}
}
\value{
Returns a vector of population sizes from \eqn{N_0,\, \ldots,\,
N_t}{N[0] to N[t]}, for integer time points.
}
\description{
Lotka-Volterra single species discrete population growth.
}
\details{
Of the form, \deqn{N_{t+1} = N_{t} + r_d N_{t}\left(1 - \alpha N_{t}\right)}
}
\examples{

library(deSolve)
# MUST use the 'euler' integration method with integer time steps
p <- c(r=1, alpha=.01)
time <- 0:10
initialN <- 10
out <- ode(y=initialN, times=time,
             func=clogistic, parms=p, method='euler')
plot(time, out[,-1], type='l')

}
\references{
Stevens. M.H.H. (2009) \emph{A Primer of Ecology with R}. Use R!
Series. Springer.
}
\seealso{
\code{\link{clogistic}}, \code{\link{dlvcomp2}},
\code{\link{lvcompg}}
}
\author{
Hank Stevens (HStevens@muohio.edu)
}
\keyword{methods}
