% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_arithmetic_seq.R
\name{generate_arithmetic_seq}
\alias{generate_arithmetic_seq}
\title{Generate an Arithmetic Sequence}
\usage{
generate_arithmetic_seq(a, d, n)
}
\arguments{
\item{a}{The first term in the sequence.}

\item{d}{The common difference.}

\item{n}{The number of terms to generate.}
}
\value{
A numeric vector representing the arithmetic sequence.
}
\description{
Generate an Arithmetic Sequence
}
\examples{
generate_arithmetic_seq(1, 2, 5) # 1, 3, 5, 7, 9
}
