% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_geometric_seq.R
\name{generate_geometric_seq}
\alias{generate_geometric_seq}
\title{Generate a Geometric Sequence}
\usage{
generate_geometric_seq(a, r, n)
}
\arguments{
\item{a}{The first term in the sequence.}

\item{r}{The common ratio.}

\item{n}{The number of terms to generate.}
}
\value{
A numeric vector representing the geometric sequence.
}
\description{
Generate a Geometric Sequence
}
\examples{
generate_geometric_seq(1, 2, 5) # 1, 2, 4, 8, 16
}
