% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_harshad.R
\name{generate_harshad}
\alias{generate_harshad}
\title{Generate Harshad Numbers}
\usage{
generate_harshad(n)
}
\arguments{
\item{n}{The number of Harshad numbers to generate.}
}
\value{
A numeric vector containing the first n Harshad numbers.
}
\description{
Generate Harshad Numbers
}
\examples{
generate_harshad(5)
}
