% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pell.R
\name{generate_pell}
\alias{generate_pell}
\title{Generate Pell Sequence}
\usage{
generate_pell(n)
}
\arguments{
\item{n}{The number of terms to generate.}
}
\value{
A numeric vector containing the first n terms of the Pell sequence.
}
\description{
Generate Pell Sequence
}
\examples{
generate_pell(5)
}
