% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_primes.R
\name{generate_primes}
\alias{generate_primes}
\title{Generate Prime Numbers up to a Given Limit}
\usage{
generate_primes(n, method = c("eratosthenes", "sundaram", "atkin"))
}
\arguments{
\item{n}{An integer, the upper limit up to which prime numbers should be generated.}

\item{method}{A character string, indicating the sieve algorithm to be used.
Options are "eratosthenes" (default), "sundaram", and "atkin".}
}
\value{
A vector of prime numbers up to the specified integer.
}
\description{
Generate a list of prime numbers up to a specified integer using one of the specified sieve algorithms.
}
\examples{
generate_primes(10)
generate_primes(10, "atkin")
generate_primes(10, "sundaram")
}
\references{
Horsley, S. (1772). The Sieve of  Eratosthenes. Being an Account of His Method of Finding
All the Prime Numbers, by the Rev. Samuel Horsley, F. R. S. Philosophical Transactions
(1683-1775), 62, 327–347. http://www.jstor.org/stable/106053

V. Ramaswami Aiyar (1934). "Sundaram's Sieve for Prime Numbers". The Mathematics Student. 2 (2): 73.

Atkin, A. O. L., & D. J. Bernstein. (2004). Prime Sieves Using Binary Quadratic Forms.
Mathematics of Computation, 73(246), 1023–1030. http://www.jstor.org/stable/4099818

Sequence A000040 in The On-Line Encyclopedia of Integer Sequences, (2010), published electronically at https://oeis.org
}
