% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_amicable.R
\name{is_amicable}
\alias{is_amicable}
\title{Check if Two Numbers are Amicable}
\usage{
is_amicable(a, b)
}
\arguments{
\item{a}{The first number.}

\item{b}{The second number.}
}
\value{
A logical value indicating whether the two numbers are amicable.
}
\description{
Check if Two Numbers are Amicable
}
\examples{
is_amicable(220, 284)
}
