% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_fibonacci.R
\name{is_fibonacci}
\alias{is_fibonacci}
\title{Check if a Number is in the Fibonacci Sequence}
\usage{
is_fibonacci(num)
}
\arguments{
\item{num}{The number to check.}
}
\value{
A logical value indicating whether the number is part of the Fibonacci sequence.
}
\description{
Check if a Number is in the Fibonacci Sequence
}
\examples{
is_fibonacci(13) # TRUE
is_fibonacci(15) # FALSE
}
