% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_happy.R
\name{is_happy}
\alias{is_happy}
\title{Check if a Number is Happy}
\usage{
is_happy(num)
}
\arguments{
\item{num}{An integer to check.}
}
\value{
A logical value indicating whether the number is happy.
}
\description{
Check if a Number is Happy
}
\examples{
is_happy(19)
}
