% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_palindrome.R
\name{is_palindrome}
\alias{is_palindrome}
\title{Check if a Number is a Palindrome}
\usage{
is_palindrome(n)
}
\arguments{
\item{n}{A numeric value to check if it's a palindrome.}
}
\value{
A logical value indicating if the number is a palindrome.
}
\description{
This function takes a number as input and returns TRUE if the number is a palindrome and FALSE otherwise.
}
\examples{
is_palindrome(121)     # TRUE
is_palindrome(123456)  # FALSE
}
