% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_perfect.R
\name{is_perfect}
\alias{is_perfect}
\title{Check if a Number is Perfect}
\usage{
is_perfect(num)
}
\arguments{
\item{num}{An integer to check.}
}
\value{
A logical value indicating whether the number is perfect.
}
\description{
Check if a Number is Perfect
}
\examples{
is_perfect(28)
}
