% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_prime.R
\name{is_prime}
\alias{is_prime}
\title{Check if a Number is Prime}
\usage{
is_prime(num)
}
\arguments{
\item{num}{Numeric, the number to check.
'}
}
\value{
Logical, TRUE if the number is prime, otherwise FALSE.
}
\description{
This function checks if a given number is prime.
}
\examples{
is_prime(7) # TRUE
is_prime(10) # FALSE
}
