% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_semi_prime.R
\name{is_semi_prime}
\alias{is_semi_prime}
\title{Check if a Number is a Semi-Prime}
\usage{
is_semi_prime(n)
}
\arguments{
\item{n}{An integer to be checked for being semi-prime.}
}
\value{
A logical value indicating whether the input number is a semi-prime.
}
\description{
This function checks if a given number is a semi-prime
}
\examples{
is_semi_prime(10)   # TRUE
is_semi_prime(12)   # FALSE
}
\references{
Sequence A000040 in The On-Line Encyclopedia of Integer Sequences, (2010), published electronically at https://oeis.org
}
