% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/next_prime.R
\name{next_prime}
\alias{next_prime}
\title{Get Next Prime After a Given Number}
\usage{
next_prime(n)
}
\arguments{
\item{n}{Numeric the reference number.}
}
\value{
Numeric the next prime after n.
}
\description{
Returns the next prime number after the given number.
}
\examples{
next_prime(7)
}
\references{
Sequence A000040 in The On-Line Encyclopedia of Integer Sequences, (2010), published electronically at https://oeis.org
}
