% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_prime.R
\name{nth_prime}
\alias{nth_prime}
\title{Get the N-th Prime Number}
\usage{
nth_prime(n)
}
\arguments{
\item{n}{An integer indicating the position of the prime number to return.}
}
\value{
The n-th prime number.
}
\description{
This function returns the n-th prime number.
}
\examples{
nth_prime(10)    # 29
nth_prime(100)   # 541
}
\references{
Sequence A002113 in The On-Line Encyclopedia of Integer Sequences, (2010), published electronically at https://oeis.org
}
