% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth_triangle_num.R
\name{nth_triangle_num}
\alias{nth_triangle_num}
\title{Get the N-th Triangular Number}
\usage{
nth_triangle_num(n)
}
\arguments{
\item{n}{The position of the triangular number.}
}
\value{
The n-th triangular number.
}
\description{
Get the N-th Triangular Number
}
\examples{
nth_triangle_num(4) # 10
nth_triangle_num(7) # 28
}
