% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_f.R
\name{plot_f}
\alias{plot_f}
\title{plot_f}
\usage{
plot_f(z)
}
\arguments{
\item{z}{numeric vector.}
}
\description{
Function to plot a time series and its periodogram, and calculate the
confidence interval of 2 times the spectral power.
}
\examples{


## white noise
plot_f(z=runif(50))

plot_f

}
\references{
J. M. Halley. Ecology, evolution and 1/f-noise. Trends in
Ecology & Evolution, 11:33-37, 1996. O. L. Petchey, A. Gonzalez, and H. B.
Wilson. Effects on population persistence: the interaction between
environmental noise colour, intra-specific competition and space.
Proceedings of the Royal Society of London Series B, 264:1841-1847, 1997. J.
E. Cohen, C. M. Newman, A. E. Cohen, O. L. Petchey, and A. Gonzalez.
Spectral mimicry: a method of synthesizing matching time series with
different Fourier spectra. Circuits, Systems and Signal Processing,
18:431-442, 1999.
}
\seealso{
[one_over_f()] to generate 1/f noise; [spec_mimic()] to rearrange
one vector, X, to mimic the spectrum of another vector; [spectrum()] for the
hard work.
}
\author{
Hank Stevens
}
\keyword{1/f}
\keyword{color-mimicry}
\keyword{noise}
\keyword{periodogram}
\keyword{spectra}
