% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime_factorization.R
\name{prime_factorization}
\alias{prime_factorization}
\title{Prime Factorization of a Number}
\usage{
prime_factorization(n)
}
\arguments{
\item{n}{An integer to be factorized into its prime factors.}
}
\value{
A named vector where names are the prime factors and the values are their powers.
}
\description{
This function returns the prime factorization of a given number.
}
\examples{
prime_factorization(60)    # 2^2 * 3^1 * 5^1
prime_factorization(100)   # 2^2 * 5^2
}
