% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer-package.R
\docType{package}
\name{primer-package}
\alias{primer}
\alias{primer-package}
\title{primer: Functions and Data for the Book, a Primer of Ecology with R}
\description{
Functions are primarily functions for systems of ordinary differential equations, difference equations, and eigenanalysis and projection of demographic matrices; data are for examples.
}
\references{
Stevens, M.H.H. (2009) \emph{A Primer of Ecology with R}
Springer, 2nd printing.
}
\seealso{
\code{\link[primer]{lvcompg}}, \code{\link[deSolve]{ode}}
}
\keyword{internal}
