% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer-package.R
\docType{data}
\name{sparrows}
\alias{sparrows}
\title{Song Sparrow Data Set}
\format{
A data frame with 36 observations on the following 3 variables.
\describe{ \item{Year}{a numeric vector} \item{Count}{a
numeric vector} \item{ObserverNumber}{a numeric vector} }
}
\source{
http://www.pwrc.usgs.gov/BBS/
}
\description{
Song Sparrow (\emph{Melospiza melodia}) counts in Darrtown, OH, USA.  From
Sauer, J. R., J.E. Hines, and J. Fallon. 2005. The North American Breeding
Bird Survey, Results and Analysis 1966--2004. Version 2005.2. USGS Patuxent
Wildlife Research Center, Laurel, MD.
}
\examples{

data(sparrows)
## maybe
str(sparrows)
plot(Count ~ Year, sparrows)

}
\references{
Stevens, M.H.H. (2009) \emph{A Primer of Ecology with R}. Use R!
Series. Springer.
}
\seealso{
\code{\link{PopSim}}
}
\keyword{datasets}
