\name{primer-package}
\alias{primer-package}
\alias{primer}
\docType{package}
\title{
Functions and data for A Primer of Ecology with R
}
\description{
 Functions are primarily functions for systems of ordinary differential equations, difference equations, and eigenanalysis and projection of demographic matrices; data are for examples.
}
\details{
\tabular{ll}{
Package: \tab primer\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2009-03-27\cr
License: \tab GPL version 2 or later\cr
LazyLoad: \tab yes\cr
}
}
\author{
Hank Stevens <HStevens@muohio.edu>
}
\references{
Stevens, M.H.H. 2009. A Primer of Ecology with R. Springer.
}

\keyword{ package }
\seealso{
\code{\link{clogistic}}, \code{\link{levins}}, \code{\link{lvcompg}}, \code{\link{predpreyLV}}, \code{\link{scheffer}, \code{\link[deSolve]{ode}} }
}

