% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer-package.R
\docType{package}
\name{primer-package}
\alias{primer-package}
\alias{primer}
\title{Functions and data for "A Primer of Ecology with R"}
\description{
Functions are primarily functions for systems of ordinary differential
equations, difference equations, and eigenanalysis and projection of
demographic matrices; data are for examples.
}
\details{
\tabular{ll}{ Package: \tab primer\cr Type: \tab Package\cr Version: \tab
1.0\cr Date: \tab 2012-05-16\cr License: \tab GPL version 2 or later\cr
LazyLoad: \tab yes\cr }
}
\references{
Stevens, M.H.H. (2009) \emph{A Primer of Ecology with R}
Springer, 2nd printing.
}
\seealso{
\code{\link[primer]{lvcompg}}, \code{\link[deSolve]{ode}}
}
\author{
Hank Stevens <HStevens@muohio.edu>
}
\keyword{package}
